chore: move to a new component
